Fix roundUpDivision type mismatches and missing clang-format guards in Legacy CUDA kernels
roundUpDivision(T, T) requires both arguments to share a type, but
several call sites mixed IndexType with std::size_t (from
Backend::getMaxGridXSize()), unsigned int (from dim3::x), or a
narrower int (from the SliceSize non-type template parameter).
Template argument deduction fails on a mismatch, so this broke as
soon as a build actually instantiated the affected code path (the
CUDA vectorProduct of AdEllpack, BiEllpack, ChunkedEllpack, Ellpack,
SlicedEllpack, and the CSR *Prepare kernels). Fixed by casting to a
common type at each call site, matching the pattern already used
elsewhere in the (now newer) TNL library.

Also wrap the AdEllpack kernel launches (Cuda4/Cuda8/Cuda16) in
// clang-format off/on, like their neighbors already were - without
the guard, clang-format mangles <<< ... >>> into < < < ... > > >,
which nvcc can't parse.

+++ include/Legacy/CSR_impl.h
@@ -1081,7 +1081,7 @@ SpMVCSRScalarPrepare( const Real* inVector, Real* outVector, const CSR< Real, De
    /* Execute kernels on device */
    for( Index grid = 0; neededThreads != 0; ++grid ) {
       if( MAX_X_DIM * threads >= neededThreads ) {
-         blocks = roundUpDivision( neededThreads, threads );
+         blocks = roundUpDivision( neededThreads, static_cast< size_t >( threads ) );
          neededThreads = 0;
       }
       else {
@@ -1111,7 +1111,7 @@ SpMVCSRVectorPrepare( const Real* inVector, Real* outVector, const CSR< Real, De
    /* Execute kernels on device */
    for( Index grid = 0; neededThreads != 0; ++grid ) {
       if( MAX_X_DIM * threads >= neededThreads ) {
-         blocks = roundUpDivision( neededThreads, threads );
+         blocks = roundUpDivision( neededThreads, static_cast< size_t >( threads ) );
          neededThreads = 0;
       }
       else {
@@ -1328,7 +1328,7 @@ SpMVCSRLightWithoutAtomicPrepare( const Real* inVector, Real* outVector, const C
    /* Execute kernels on device */
    for( Index grid = 0; neededThreads != 0; ++grid ) {
       if( MAX_X_DIM * threads >= neededThreads ) {
-         blocks = roundUpDivision( neededThreads, threads );
+         blocks = roundUpDivision( neededThreads, static_cast< size_t >( threads ) );
          neededThreads = 0;
       }
       else {
@@ -1506,7 +1506,7 @@ SpMVCSRMultiVectorPrepare( const Real* inVector, Real* outVector, const CSR< Rea
    /* Execute kernels on device */
    for( Index grid = 0; neededThreads != 0; ++grid ) {
       if( MAX_X_DIM * threads >= neededThreads ) {
-         blocks = roundUpDivision( neededThreads, threads );
+         blocks = roundUpDivision( neededThreads, static_cast< size_t >( threads ) );
          neededThreads = 0;
       }
       else {
@@ -1660,7 +1660,7 @@ SpMVCSRAdaptivePrepare( const Real* inVector, Real* outVector, const CSR< Real, 
    // Execute kernels on device
    for( Index grid = 0; neededThreads != 0; ++grid ) {
       if( MAX_X_DIM * threads >= neededThreads ) {
-         blocks = roundUpDivision( neededThreads, threads );
+         blocks = roundUpDivision( neededThreads, static_cast< size_t >( threads ) );
          neededThreads = 0;
       }
       else {
